+++ src/main/kotlin/com/github/koooooo7/robin/db/WeekDay.kt
@@ -1,0 +1,11 @@
+package com.github.koooooo7.robin.db
+
+enum class WeekDay(val abbreviation: String, val idx: Int) {
+    Mon("Mon", 0),
+    Tue("Tue", 1),
+    Wed("Wed", 2),
+    Thu("Thu", 3),
+    Fri("Fri", 4),
+    Sat("Sat", 5),
+    Sun("Sun", 6),
+}