+++ src/main/kotlin/com/github/koooooo7/robin/worker/RingerWorker.kt
@@ -1,6 +1,7 @@
 package com.github.koooooo7.robin.worker
 
 import com.github.koooooo7.robin.db.RingerSettings
+import com.github.koooooo7.robin.db.WeekDay
 import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.ui.Messages
 import com.jetbrains.rd.util.ConcurrentHashMap
@@ -12,29 +13,20 @@ import java.util.*
 
 class RingerWorker {
     private var setup = false
-    private val dayOfWeekIndexMap = mapOf(
-        "Mon" to 0,
-        "Tue" to 1,
-        "Wed" to 2,
-        "Thu" to 3,
-        "Fri" to 4,
-        "Sat" to 5,
-        "Sun" to 6,
-    )
 
-    private var currentDayCacheDay = dayOfWeekIndexMap["Mon"]
+    private var currentDayCacheDay = WeekDay.Mon
     private val runtimeRingedCache = ConcurrentHashMap<Int, Int>()
 
     fun start(ringerSettings: RingerSettings) {
 
-        Thread {
+        val t = Thread {
             while (true) {
                 Thread.sleep(10_000)
-                val ringers = ringerSettings.getRingers()
+                val ringers = ringerSettings.getRingers().values
                 val today = LocalDate.now()
                 val currentDay = today.dayOfWeek.getDisplayName(TextStyle.SHORT, Locale.getDefault())
-                if (currentDayCacheDay != dayOfWeekIndexMap[currentDay]!!) {
-                    currentDayCacheDay = dayOfWeekIndexMap[currentDay]!!
+                if (currentDayCacheDay != WeekDay.valueOf(currentDay)) {
+                    currentDayCacheDay = WeekDay.valueOf(currentDay)
                     runtimeRingedCache.clear()
                 }
 
@@ -45,7 +37,7 @@ class RingerWorker {
                     val selectedHour = ringer.selectedHour
                     val selectedMinute = ringer.selectedMinute
                     val selectedWeekend = ringer.weekend
-                    if (!selectedWeekend[dayOfWeekIndexMap[currentDay]!!]) {
+                    if (!selectedWeekend[WeekDay.valueOf(currentDay).idx]) {
                         continue
                     }
 
@@ -83,15 +75,17 @@ class RingerWorker {
                         runtimeRingedCache[currentHour] = currentMinute
                     }
 
-                    val title = "[$currentDay $currentHour:$currentMinute] Robin ring :>"
+                    val title = "[$currentDay ${String.format("%02d", currentHour)}:${String.format("%02d", currentMinute)}] Robin ring :>"
 
                     ApplicationManager.getApplication().invokeLater {
                         Messages.showMessageDialog(ringer.description, title, Messages.getInformationIcon())
                     }
                 }
             }
+        }
 
-        }.start()
+        t.isDaemon = true
+        t.start()
 
     }