+++ CHANGELOG.md
@@ -2,9 +2,12 @@
 
 # robin Changelog
 
-## [0.0.1]
+## [0.0.2]
 
-### Added
+- Support remove single ringer.
+- Update the duplicated check for ringers.
+
+## [0.0.1]
 
 - Support add/reset multi ringers.
 - Support set robin time and the work day of week.